@@ -1,29 +1,22 @@
 {
   "name": "tdi",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "homepage": "http://www.kavispires.com/tdi/",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^5.0.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/lodash": "^4.14.191",
-    "@types/node": "^17.0.41",
-    "@types/react": "^17.0.45",
-    "@types/react-dom": "^17.0.17",
-    "@types/react-router-dom": "5.3.3",
-    "antd": "^5.1.5",
-    "clsx": "^1.2.1",
-    "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "6.2.1",
-    "react-scripts": "4.0.3",
-    "react-use": "^17.4.0",
-    "sass": "^1.57.1",
-    "typescript": "^4.9.4",
-    "web-vitals": "^1.0.1"
+    "@testing-library/jest-dom": "^6.4.1",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/node": "^20.11.16",
+    "@types/react": "^18.2.52",
+    "@types/react-dom": "^18.2.18",
+    "postcss-normalize": "^10.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "sass": "^1.70.0",
+    "typescript": "^5.3.3",
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -76,9 +69,9 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.2.3",
+    "gh-pages": "^6.1.1",
     "husky": "^8.0.3",
     "lint-staged": "^13.1.0",
-    "prettier": "^2.8.3"
+    "prettier": "^3.2.4"
   }
 }
